+++ packages/shared/src/index.ts
@@ -84,6 +84,7 @@ export * from './recipe';
 export * from './bodyweight';
 export * from './weight-goal';
 export * from './muscle-balance';
+export * from './goal-letter';
 export * from './fuel-tank';
 export * from './run-ghost';
 export * from './running';